Placed the cards into the TotalsRow
