+++ dashboard/src/components/ContentWrapper/TotalsRow/TotalsRowContainer.jsx
@@ -2,10 +2,13 @@ import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import Paper from "@material-ui/core/Paper";
 import Grid from "@material-ui/core/Grid";
+import SmallCard from "../../SmallCard/SmallCard";
+import LineGraph from "../../AreaGraph/AreaGraph";
 
 const useStyles = makeStyles((theme) => ({
 	root: {
 		flexGrow: 1,
+		marginBottom: 25
 	},
 	paper: {
 		padding: theme.spacing(2),
@@ -14,26 +17,54 @@ const useStyles = makeStyles((theme) => ({
 	},
 }));
 
+const chart = {
+	width: 225,
+	height: 150,
+};
+
 export default function TotalsRowContainer() {
 	const classes = useStyles();
 
 	return (
 		<div className={classes.root}>
-			TotalsRow Container
-				<Grid container spacing={3}>
-					<Grid item xs={12} md={6} lg={3} spacing={1}>
-						<Paper className={classes.paper}> Users Total</Paper>
-					</Grid>
-					<Grid item xs={12} md={6} lg={3} spacing={1}>
-						<Paper className={classes.paper}> Posts Total</Paper>
-					</Grid>
-					<Grid item xs={12} md={6} lg={3} spacing={1}>
-						<Paper className={classes.paper}> Card3</Paper>
-					</Grid>
-					<Grid item xs={12} md={6} lg={3} spacing={1}>
-						<Paper className={classes.paper}> Card4</Paper>
-					</Grid>
+			<Grid container spacing={3}>
+				<Grid item xs={12} md={4} lg={3} spacing={1}>
+					<SmallCard
+						backgroundColor="red"
+						title="Usuarios"
+						total="123"
+						chart={chart}
+					/>
+					<LineGraph />
 				</Grid>
+				<Grid item xs={12} md={4} lg={3} spacing={1}>
+					<SmallCard
+						backgroundColor="green"
+						title="posts"
+						total="123"
+						chart={chart}
+					/>
+					<LineGraph />
+				</Grid>
+				<Grid item xs={12} md={4} lg={3} spacing={1}>
+					<SmallCard
+						backgroundColor="blue"
+						title="Vehicles"
+						total="123"
+						chart={chart}
+					/>
+					<LineGraph />
+				</Grid>
+				<Grid item xs={12} md={4} lg={3} spacing={1}>
+					<SmallCard
+						backgroundColor="orange"
+						title="Parts"
+						total="123"
+						chart={chart}
+					/>
+					<LineGraph />
+				</Grid>
+			</Grid>
 		</div>
 	);
 }
